Factorize code from OAuth1Swift & OAuth2Swift to new file OAuthSwift Fix #26 by making credential getter public
