+++ OAuthSwiftDemo/AppDelegate.swift
@@ -53,16 +53,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UIWebViewDelegate {
     func application(application: UIApplication, openURL url: NSURL, sourceApplication: String?, annotation: AnyObject?) -> Bool {
         println(url)
         if (url.host == "oauth-callback") {
-            if (url.path!.hasPrefix("/twitter") || url.path!.hasPrefix("/flickr") || url.path!.hasPrefix("/fitbit")
-             || url.path!.hasPrefix("/withings") || url.path!.hasPrefix("/linkedin") || url.path!.hasPrefix("/bitbucket") || url.path!.hasPrefix("/smugmug") || url.path!.hasPrefix("/intuit") ) {
-                OAuth1Swift.handleOpenURL(url)
-            }
-            if ( url.path!.hasPrefix("/github" ) || url.path!.hasPrefix("/instagram" ) || url.path!.hasPrefix("/foursquare") || url.path!.hasPrefix("/dropbox") || url.path!.hasPrefix("/dribbble") || url.path!.hasPrefix("/salesforce") || url.path!.hasPrefix("/google") || url.path!.hasPrefix("/linkedin2")) {
-                OAuth2Swift.handleOpenURL(url)
-            }
+            OAuthSwift.handleOpenURL(url)
         } else {
             // Google provider is the only one wuth your.bundle.id url schema.
-            OAuth2Swift.handleOpenURL(url)
+            OAuthSwift.handleOpenURL(url)
         }
         return true
     }
