+++ OAuthSwiftOSXDemo/AppDelegate.swift
@@ -12,11 +12,19 @@ import OAuthSwiftOSX
 @NSApplicationMain
 class AppDelegate: NSObject, NSApplicationDelegate {
 
-
-
     func applicationDidFinishLaunching(aNotification: NSNotification) {
-        
+        // Register to url events
         NSAppleEventManager.sharedAppleEventManager().setEventHandler(self, andSelector:"handleGetURLEvent:withReplyEvent:", forEventClass: AEEventClass(kInternetEventClass), andEventID: AEEventID(kAEGetURL))
+        
+        // Code to check configuration project
+        if let urlTypes = NSBundle.mainBundle().infoDictionary!["CFBundleURLTypes"] as? Array<AnyObject>,
+            urlType = urlTypes.first as? [String:AnyObject],
+            urlSchemes = urlType["CFBundleURLSchemes"] as? [String] {
+                if !contains(urlSchemes, "oauth-swift") {
+                    println("oauth-swift no registered to url schemes")
+                }
+        }
+
     }
 
     func applicationWillTerminate(aNotification: NSNotification) {
@@ -32,20 +40,9 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     
     internal func applicationHandleOpenURL(url: NSURL) {
         println(url)
-        if (url.host == "oauth-callback") {
-            if (url.path!.hasPrefix("/twitter") || url.path!.hasPrefix("/flickr") || url.path!.hasPrefix("/fitbit")
-                || url.path!.hasPrefix("/withings") || url.path!.hasPrefix("/linkedin") || url.path!.hasPrefix("/bitbucket") || url.path!.hasPrefix("/smugmug") || url.path!.hasPrefix("/intuit") || url.path!.hasPrefix("/trello") ) {
-                    OAuth1Swift.handleOpenURL(url)
-            }
-            if ( url.path!.hasPrefix("/github" ) || url.path!.hasPrefix("/instagram" ) || url.path!.hasPrefix("/foursquare") || url.path!.hasPrefix("/dropbox") || url.path!.hasPrefix("/dribbble") || url.path!.hasPrefix("/salesforce") || url.path!.hasPrefix("/google") || url.path!.hasPrefix("/linkedin2")) {
-                OAuth2Swift.handleOpenURL(url)
-            }
-        } else {
-            // Google provider is the only one wuth your.bundle.id url schema.
-            OAuth2Swift.handleOpenURL(url)
-        }
+        OAuthSwift.handleOpenURL(url)
     }
-    
+
     class var sharedInstance: AppDelegate {
         return NSApplication.sharedApplication().delegate as! AppDelegate
     }
